+++ src/core/io/glaxnimate/glaxnimate_importer.cpp
@@ -24,9 +24,6 @@ bool io::glaxnimate::GlaxnimateFormat::on_open ( QIODevice& file, const QString&
 
     QJsonObject top_level = jdoc.object();
 
-
-    document->metadata() = top_level["metadata"].toObject().toVariantMap();
-
     if ( !top_level["animation"].isObject() )
     {
         error(tr("Missing animation object"));
@@ -38,6 +35,7 @@ bool io::glaxnimate::GlaxnimateFormat::on_open ( QIODevice& file, const QString&
     if ( document_version > format_version )
         warning(tr("Opening a file from a newer version of Glaxnimate"));
 
+    state.load_metadata(document, top_level);
     state.load_object(document->assets(), top_level[document_version < 3 ? "defs" : "assets"].toObject());
     state.load_object(document->main(), top_level["animation"].toObject());
     state.resolve();