+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * Fixed loading layers with a masked parent from lottie
     * Fixed exporting polygons to lottie
     * Added support for loading (old) lottie v 4.0.0
+    * Added support for common metadata (author/description) across formats
 * UI:
     * Added simplified Chinese translation
     * Added British English translation